# src/controllers/postController.js
const PostModel = require('../models/Post')

// GET /posts/search
exports.searchPosts = async (req, res) => {
  const { term: searchTerm } = req.query;
  if (!searchTerm || !searchTerm.trim()) return res.status(400).json({
    error: 'Termo(s) de busca necessário(s)'
  });
  try {
    const keywords = searchTerm.split(' ').filter(Boolean);
    const results = await PostModel.searchPosts(keywords);
    res.status(200).json(results);
  } catch (error) {
    res.status(500).json({
      error: 'Erro na busca'
    });
  }
}

// GET /posts
exports.getAllPosts = async (req, res) => {
  try {
    const posts = await PostModel.getAllPosts();
    res.status(200).json(posts);
  } catch (error) {
    res.status(500).json({
      error: 'Erro ao buscar todos os posts'
    });
  }
}

// POST /posts
exports.createPost = async (req, res) => {
  const { title, content, author } = req.body;

  try {
    const newPost = await PostModel.createPost({ title, content, author });
    res.status(201).json(newPost);
  } catch (error) {
    res.status(500).json({
      error: 'Erro ao criar post'
    });
  }
}

// GET /posts/:id
exports.getPostById = async (req, res) => {
  const id = req.params.id

  try {
    const postById = await PostModel.getPostById(id);

    if (!postById) return res.status(404).json({
      error: 'Post não encontrado'
    });

    res.status(200).json(postById);
  } catch (error) {
    res.status(500).json({
      error: 'Erro ao buscar post'
    });
  }
}

// PUT /posts/:id
exports.updatePost = async (req, res) => {
  const id = req.params.id
  const { title, content } = req.body

  try {
    const updatedPost = await PostModel.updatePost(id, { title, content });

    if (!updatedPost) return res.status(404).json({
      error: 'Post não encontrado'
    });

    res.status(200).json(updatedPost);
  } catch (error) {
    res.status(500).json({
      error: 'Erro ao atualizar post'
    });
  }
}

// DELETE /posts/:id
exports.deletePostById = async (req, res) => {
  const id = req.params.id

  try {
    const post = await PostModel.deletePostById(id);

    if (!post) return res.status(404).json({
      error: 'Post não encontrado'
    });

    res.status(204).send();
  } catch (error) {
    res.status(500).json({
      error: 'Erro ao deletar post'
    });
  }
}